pyproject.toml
[project]
name = "np"
version = "0.1.0"
description = "forum signatures with last.fm"
authors = [
    {name = "Niko Shens"}
]
requires-python = ">=3.12"
dependencies = [
    "flask (>=3.1.3,<4.0.0)",
    "waitress (>=3.0.2,<4.0.0)",
    "pylast (>=7.0.2,<8.0.0)",
    "pillow (>=12.2.0,<13.0.0)",
    "requests (>=2.33.1,<3.0.0)"
]

[dependency-groups]
dev = [
    "black (>=26.3.1,<27.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
